% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mail.R
\name{get_mailbox}
\alias{get_mailbox}
\title{Retrieve the mailbox object}
\usage{
get_mailbox(path, type)
}
\arguments{
\item{path}{The path to the mailbox}

\item{type}{The format in which the mailbox is stored}
}
\value{
A mailbox object
}
\description{
Retrieve the mailbox object
}
