% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{get_messages}
\alias{get_messages}
\title{Parse all messages from the given mail_object and return a data.frame}
\usage{
get_messages(mail_object)
}
\arguments{
\item{mail_object}{A mail_object as read in by read_mail}
}
\value{
A data.frame containing all the fields from the message including the body
}
\description{
Parse all messages from the given mail_object and return a data.frame
}
