% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailman.R
\docType{package}
\name{mailman-package}
\alias{mailman}
\alias{mailman-package}
\title{mailman: a wrapper around the python mailbox module}
\description{
The python mailbox module provides functionality to manipulate mailboxes in various formats.
This package is a wrapper around the python module and exposes a number of functions from it.
}
\author{
\strong{Maintainer}: Adnan Fiaz \email{afiaz@mango-solutions.com}

}
