% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mail.R
\name{read_messages}
\alias{read_messages}
\title{Read e-mail messages from an on-disk mailbox}
\usage{
read_messages(path, type = c("mbox", "MailDir", "MH", "Babyl", "MMDF"))
}
\arguments{
\item{path}{The path to the mailbox}

\item{type}{The format in which the mailbox is stored}
}
\value{
a data.frame with the headers and body of the messages
}
\description{
Read e-mail messages from an on-disk mailbox
}
